+++ public/app/core/components/NavBar/NavBarItemMenuTrigger.tsx
@@ -1,27 +1,16 @@
-import React, {
-  createContext,
-  HTMLAttributes,
-  Key,
-  ReactElement,
-  useContext,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+import React, { ReactElement, useState } from 'react';
 import { css, cx } from '@emotion/css';
 import { Icon, IconName, Link, useTheme2 } from '@grafana/ui';
-import { GrafanaTheme2, NavMenuItemType, NavModelItem } from '@grafana/data';
-import { MenuTriggerProps, SpectrumMenuProps } from '@react-types/menu';
+import { GrafanaTheme2, NavModelItem } from '@grafana/data';
+import { MenuTriggerProps } from '@react-types/menu';
 import { useMenuTriggerState } from '@react-stately/menu';
-import { useMenu, useMenuItem, useMenuTrigger } from '@react-aria/menu';
-import { useFocus, useFocusVisible, useFocusWithin, useHover, useKeyboard } from '@react-aria/interactions';
+import { useMenuTrigger } from '@react-aria/menu';
+import { useFocusVisible, useFocusWithin, useHover, useKeyboard } from '@react-aria/interactions';
 import { useButton } from '@react-aria/button';
-import { TreeState, useTreeState } from '@react-stately/tree';
 import { DismissButton, useOverlay } from '@react-aria/overlays';
 import { FocusScope } from '@react-aria/focus';
-import { mergeProps } from '@react-aria/utils';
-import { Node } from '@react-types/shared';
-import { getNavModelItemKey } from './utils';
+
+import { NavBarItemMenuContext } from './context';
 
 export interface NavBarItemMenuTriggerProps extends MenuTriggerProps {
   children: ReactElement;
@@ -218,234 +207,3 @@ const getStyles = (theme: GrafanaTheme2, isActive?: boolean) => ({
     }
   `,
 });
-
-export interface NavBarItemMenuContextProps {
-  enableAllItems: boolean;
-  onClose: () => void;
-  menuProps?: HTMLAttributes<HTMLElement>;
-}
-
-export const NavBarItemMenuContext = createContext<NavBarItemMenuContextProps>({
-  enableAllItems: false,
-  onClose: () => undefined,
-});
-
-export function useNavBarItemMenuContext(): NavBarItemMenuContextProps {
-  return useContext(NavBarItemMenuContext);
-}
-
-export interface NavBarItemMenuProps extends SpectrumMenuProps<NavModelItem> {
-  adjustHeightForBorder: boolean;
-  reverseMenuDirection?: boolean;
-}
-
-export function NavBarItemMenu(props: NavBarItemMenuProps): ReactElement | null {
-  const { reverseMenuDirection, adjustHeightForBorder, disabledKeys, onAction, ...rest } = props;
-  const contextProps = useNavBarItemMenuContext();
-  const completeProps = {
-    ...mergeProps(contextProps, rest),
-  };
-  const { enableAllItems, menuProps: contextMenuProps = {} } = contextProps;
-  const theme = useTheme2();
-  const styles = getNavBarItemMenuStyles(theme, adjustHeightForBorder, reverseMenuDirection);
-  const state = useTreeState<NavModelItem>({ ...rest, disabledKeys });
-  const ref = useRef(null);
-  const { menuProps } = useMenu(completeProps, { ...state }, ref);
-  const allItems = [...state.collection];
-  const items = allItems.filter((item) => item.value.menuItemType === NavMenuItemType.Item);
-  const section = allItems.find((item) => item.value.menuItemType === NavMenuItemType.Section);
-
-  useEffect(() => {
-    if (enableAllItems && !state.selectionManager.isFocused) {
-      state.selectionManager.setFocusedKey(section?.key ?? '');
-      state.selectionManager.setFocused(true);
-    } else if (!enableAllItems && state.selectionManager.isFocused) {
-      state.selectionManager.setFocused(false);
-      state.selectionManager.clearSelection();
-    }
-  }, [enableAllItems, state.selectionManager, reverseMenuDirection, section?.key]);
-
-  if (!section) {
-    return null;
-  }
-
-  const menuSubTitle = section.value.subTitle;
-
-  const sectionComponent = (
-    <NavBarItemMenuItem
-      key={section.key}
-      item={section}
-      state={state}
-      onAction={onAction}
-      reverseDirection={reverseMenuDirection}
-    />
-  );
-
-  const subTitleComponent = (
-    <li key={menuSubTitle} className={styles.menuItem}>
-      <div className={styles.subtitle}>{menuSubTitle}</div>
-    </li>
-  );
-
-  return (
-    <ul
-      className={`${styles.menu} navbar-dropdown`}
-      ref={ref}
-      {...mergeProps(menuProps, contextMenuProps)}
-      tabIndex={enableAllItems ? 0 : -1}
-    >
-      {!reverseMenuDirection ? sectionComponent : null}
-      {menuSubTitle && reverseMenuDirection ? subTitleComponent : null}
-      {items.map((item, index) => {
-        return (
-          <NavBarItemMenuItem
-            key={getNavModelItemKey(item.value)}
-            item={item}
-            state={state}
-            onAction={onAction}
-            reverseDirection={reverseMenuDirection}
-          />
-        );
-      })}
-      {reverseMenuDirection ? sectionComponent : null}
-      {menuSubTitle && !reverseMenuDirection ? subTitleComponent : null}
-    </ul>
-  );
-}
-
-function getNavBarItemMenuStyles(
-  theme: GrafanaTheme2,
-  adjustHeightForBorder: boolean,
-  reverseDirection?: boolean,
-  isFocused?: boolean
-) {
-  return {
-    menu: css`
-      background-color: ${theme.colors.background.primary};
-      border: 1px solid ${theme.components.panel.borderColor};
-      bottom: ${reverseDirection ? 0 : 'auto'};
-      box-shadow: ${theme.shadows.z3};
-      display: flex;
-      flex-direction: column;
-      left: 100%;
-      list-style: none;
-      min-width: 140px;
-      position: absolute;
-      top: ${reverseDirection ? 'auto' : 0};
-      transition: ${theme.transitions.create('opacity')};
-      z-index: ${theme.zIndex.sidemenu};
-      list-style: none;
-    `,
-    header: css`
-      background-color: ${theme.colors.background.secondary};
-      color: ${theme.colors.text.primary};
-      height: ${theme.components.sidemenu.width - (adjustHeightForBorder ? 2 : 1)}px;
-      font-size: ${theme.typography.h4.fontSize};
-      font-weight: ${theme.typography.h4.fontWeight};
-      padding: ${theme.spacing(1)} ${theme.spacing(2)};
-      white-space: nowrap;
-      width: 100%;
-    `,
-    menuItem: css`
-      background-color: ${isFocused ? theme.colors.action.hover : 'transparent'};
-      color: ${isFocused ? 'white' : theme.colors.text.primary};
-
-      &:focus-visible {
-        background-color: ${theme.colors.action.hover};
-        box-shadow: none;
-        color: ${theme.colors.text.primary};
-        outline: 2px solid ${theme.colors.primary.main};
-        // Need to add condition, header is 0, otherwise -2
-        outline-offset: -0px;
-        transition: none;
-      }
-    `,
-    subtitle: css`
-      border-${reverseDirection ? 'bottom' : 'top'}: 1px solid ${theme.colors.border.weak};
-      color: ${theme.colors.text.secondary};
-      font-size: ${theme.typography.bodySmall.fontSize};
-      font-weight: ${theme.typography.bodySmall.fontWeight};
-      padding: ${theme.spacing(1)} ${theme.spacing(2)} ${theme.spacing(1)};
-      white-space: nowrap;
-    `,
-  };
-}
-
-export interface NavBarItemMenuItemProps {
-  item: Node<NavModelItem>;
-  state: TreeState<NavModelItem>;
-  reverseDirection?: boolean;
-  onAction?: (key: Key) => void;
-}
-
-export function NavBarItemMenuItem({
-  item,
-  state,
-  onAction,
-  reverseDirection = false,
-}: NavBarItemMenuItemProps): ReactElement {
-  const { onClose } = useNavBarItemMenuContext();
-  const { key, rendered } = item;
-  const ref = useRef<HTMLLIElement>(null);
-  const isDisabled = state.disabledKeys.has(key);
-
-  // style to the focused menu item
-  const [isFocused, setFocused] = useState(false);
-  const { focusProps } = useFocus({ onFocusChange: setFocused, isDisabled });
-  const theme = useTheme2();
-  const styles = getStylesMenuItem(theme, isFocused, reverseDirection);
-
-  let { menuItemProps } = useMenuItem(
-    {
-      isDisabled,
-      'aria-label': item['aria-label'],
-      key,
-      onClose,
-      closeOnSelect: true,
-      onAction,
-    },
-    state,
-    ref
-  );
-
-  const onKeyDown = (e: React.KeyboardEvent) => {
-    switch (e.key) {
-      case 'Enter':
-      case ' ':
-        // Stop propagation, unless it would already be handled by useKeyboard.
-        if (!('continuePropagation' in e)) {
-          e.stopPropagation();
-        }
-        e.preventDefault();
-        // Alert: Hacky way to go to link
-        e.currentTarget?.querySelector('a')?.click();
-        e.currentTarget?.querySelector('button')?.click();
-        break;
-    }
-  };
-
-  return (
-    <li {...mergeProps(menuItemProps, focusProps)} onKeyDown={onKeyDown} ref={ref} className={styles.menuItem}>
-      {rendered}
-    </li>
-  );
-}
-
-function getStylesMenuItem(theme: GrafanaTheme2, isFocused: boolean, reverseDirection: boolean) {
-  return {
-    menuItem: css`
-      background-color: ${isFocused ? theme.colors.action.hover : 'transparent'};
-      color: ${isFocused ? 'white' : theme.colors.text.primary};
-
-      &:focus-visible {
-        background-color: ${theme.colors.action.hover};
-        box-shadow: none;
-        color: ${theme.colors.text.primary};
-        outline: 2px solid ${theme.colors.primary.main};
-        // Need to add condition, header is 0, otherwise -2
-        outline-offset: -0px;
-        transition: none;
-      }
-    `,
-  };
-}
