+++ public/app/core/components/NavBar/context.tsx
@@ -1,0 +1,16 @@
+import { createContext, HTMLAttributes, useContext } from 'react';
+
+export interface NavBarItemMenuContextProps {
+  enableAllItems: boolean;
+  onClose: () => void;
+  menuProps?: HTMLAttributes<HTMLElement>;
+}
+
+export const NavBarItemMenuContext = createContext<NavBarItemMenuContextProps>({
+  enableAllItems: false,
+  onClose: () => undefined,
+});
+
+export function useNavBarItemMenuContext(): NavBarItemMenuContextProps {
+  return useContext(NavBarItemMenuContext);
+}
